update geoinfo
update geoinfo to prepare and save all known stiltstations with country information.
Switch to : if it is an icos station, use the icos lat/lon
otherwise the stilt lat/lon.

ICOS lat/lon is more precis. Importat for example Zugspitze...with STILT coordiantes  the station belongs to Austria, with ICOS to Germany.